Moved commands to a separate file
